+++ CHANGELOG.md
@@ -19,7 +19,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Removed
 
+- Removed `@ForceJson` annotation.
+
 ### Fixed
+
+- Fixed bug, where Media with Content Type Application/Json cannot be downloaded.
 
 ### Security